+++ twine-components/src/thermal/tank.rs
@@ -5,8 +5,9 @@ use twine_core::{
             CvProvider, DensityProvider, EnthalpyProvider, FluidPropertyError, FluidPropertyModel,
             TemperatureProvider,
         },
-        units::{PositiveMassRate, TemperatureDifference, TimeDerivativeOf},
+        units::{PositiveMassRate, TemperatureDifference},
     },
+    transient::TimeDerivativeOf,
     Component,
 };
 use uom::si::f64::{Area, HeatTransfer, Power, ThermodynamicTemperature, Volume};
