+++ twine-core/src/lib.rs
@@ -1,8 +1,8 @@
 mod component;
 pub mod graph;
-pub mod simulation;
 pub mod solve;
 pub mod thermo;
+pub mod transient;
 mod twine;
 
 pub use component::Component;
