--- twine-core/src/simulation/test_utils.rs
@@ -1,4 +1,0 @@
-mod moving_point;
-
-#[allow(unused_imports)]
-pub(crate) use moving_point::{MovingPoint, PointInput, PointOutput};
