+++ twine-core/src/transient/test_utils.rs
@@ -1,0 +1,3 @@
+mod moving_point;
+
+pub(crate) use moving_point::{MovingPoint, PointInput};
