+++ twine-core/src/transient/traits/state_integrator.rs
